refactor: rename infraestructure to application
