+++ app/adapters/repositories/strategy/proxies/proxy_write.py
@@ -1,6 +1,6 @@
 from typing import List, Any
 from app.config.messages import ERR_SECURE_PROXY as ERR_MSG
-from app.infraestructure.exceptions.proxy import OverrideError
+from app.application.exceptions.proxy import OverrideError
 
 
 class WritableProxy:
