+++ app/application/screens/templates/civilizations_start.txt
