+++ tests/unitary/test_infraestructure_logging_custom_logger.py
@@ -2,7 +2,7 @@ import init  # noqa: F401
 from unittest import TestCase
 from os import environ
 from time import sleep
-from app.infraestructure.logging import logger, DEBUG
+from app.application.logging import logger, DEBUG
 
 
 class TestInfraestructureLoggingCustomLogger(TestCase):
